Decouple LimitingExecutor into separate class
